{histogram,predictor}_enc: quiet int -> float warnings
after:
a19a25bb Replace doubles by floats in lossless misc cost estimations.

+++ src/enc/predictor_enc.c
@@ -35,7 +35,7 @@ static float PredictionCostSpatial(const int counts[256], int weight_0,
                                    float exp_val) {
   const int significant_symbols = 256 >> 4;
   const float exp_decay_factor = 0.6f;
-  float bits = weight_0 * counts[0];
+  float bits = (float)weight_0 * counts[0];
   int i;
   for (i = 1; i < significant_symbols; ++i) {
     bits += exp_val * (counts[i] + counts[256 - i]);
